requirements.txt
annotated-types==0.7.0
anyio==4.9.0
beautifulsoup4==4.13.4
certifi==2025.6.15
click==8.2.1
colorama==0.4.6
distro==1.9.0
fastapi==0.115.14
groq==0.30.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
jiter==0.10.0
joblib==1.5.1
numpy==2.3.1
openai==1.92.2
pydantic==1.10.22
pydantic_core==2.33.2
python-dotenv==1.1.1
redis==6.2.0
scikit-learn==1.7.0
scipy==1.16.0
sniffio==1.3.1
soupsieve==2.7
starlette==0.46.2
structlog==25.4.0
threadpoolctl==3.6.0
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.0
uvicorn==0.34.3
fastapi==0.110.*
uvicorn[standard]==0.29.*
groq==0.9.*
python-dotenv==1.0.*

